Use search component in new viz wizard
